ENH: Update NIRFast-meshing + http protocol
(1) We updated the version of NIRFAST-meshing found on github at
https://github.com/nirfast-admin/nirfast-meshing-dev/tree/NIRView
that adds the implementation fo the fcsv reader for fiducials from
Slicer, as well as the Transform Matrix correction that was not taken
into account when calculating the offset.

(2) Also updated the git protocols to http to avoid a problem with git
on Windows.