ENH: Write VTK Dataset in file system
There was no need to write the polydata mesh (not made for tetrahedral
mesh) in NIRView. We know use the old implementation fo nirfast2mesh
to write the VTK dataset (unstructured grid) on the file system.

This file will be used in Mesh2Image to extract its point data information
in volunes that we can load in Slicer.

[WIP]: in the future we will need to create a new module calling
nirfast2mesh so that the nirfast mesh can first be processed.