BUG: Reset DisplayNode to None in Mesh2Image
If the copied node (cropped volume) was being rendered as a volume
in Slicer, them it had another display node that did not disappear
when setting it to None after copying the volume. We have for some
reason to set it to None twice more : after the node is added to
the scene, and after it is added to a combobox.