ENH: Duplicate Shade checkbox in Volume Rendering module
In case of optical properties we are not interested in using
Shaders (no need for photo-realistic rendering, bad rendering
if cropped). For inexperienced user, the process of unchecking
this box is finnacky, since it is located in the Advanced box
of the Volume property tab of the Advanced collapsible menu.

To make this easier, we create a duplicate of this checkbox
in the display menu (top of the module menu).

To access the checkbox we needed the ctkVTKVolumePropertyWidget
within the qMRMLVolumePropertyNodeWidget of the main module widget.
To do so, we needed to add the qSlicerVolumeRenderingModuleWidgets
target to link against the main application. This required the use
of the slicerMacroBuildApplication macro which was not using its
INCLUDE_DIRECTORIES argument correctly. We therefore update 3DSlicer
version to include commit 52b12fa which fixes that issue.
-