BUG: Fix addpath bug on windows
The issue was due to pn.path that could be a system path,
in which case writting permissions would be denied.

By using userpath we ensure that this file will be writable.

As a safeguard, we'll display an error message explaining
how to manually add the path in MatLab if it fails.